Use new RK3576 U-Boot branch
Most of the RK3576 support got merged into mainline U-Boot, so I thought
it was a good time to recreate our branch based on v2025.07-rc1. While I
was at it, I also renamed the branch to something more accurate. I
avoided calling it just "rk3576" because people's repos might already
have that ref as a tag, which was the case for me for some reason.

Things that are in this branch but aren't in mainline yet (AFAIK):
- the USB-C TCPM things
- Kwiboo's SD card boost hack, required for SD card booting
- the Sige5 device tree
- the fusb302 node within said device tree

Signed-off-by: Nicolas Frattaroli <nicolas.frattaroli@collabora.com>