Add support for device tree overlays to RK3588 image
Often, a board is extended with additional hardware, such as camera
modules. Some modules can be detected, but not all of them. For those
that cannot be detected, we can add a list of required device tree
overlays to the U-Boot menu configuration. Configure the device tree
overlay directory correctly and add an example of how this could be
achieved.

Signed-off-by: Michael Riesch <michael.riesch@collabora.com>